USE the right DATATYPES! dont use doubles - use Evas_Coord.
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@9923 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ misc/engage/src/window.c
@@ -124,7 +124,7 @@ od_window_init()
   Ecore_X_Display *dsp;
   Screen         *scr;
   int             def;
-  double          res_x, res_y;
+  int             res_x, res_y;
   Evas_Object    *o;
 
   // determine the desktop size
